pyproject.toml
[tool.poetry]
name = "fastapiprojectforshinhan"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.14"
fastapi = "^0.128.0"
uvicorn = "^0.40.0"
structlog = "^25.5.0"
black = "^25.12.0"
orjson = "^3.11.5"
sqlalchemy = "^2.0.45"
pydantic-settings = "^2.12.0"
supabase = "^2.27.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
httpx = "^0.28"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
